refactor: move header component to common dir
