+++ src/components/Common/Header/GoBackBtn.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useRouter } from 'next/router';
 import GoBackSvgIcon from 'public/assets/svg/chevron-left.svg';
-import { SvgIcon } from '../SvgStyle.styled';
+import { SvgIcon } from '../../SvgStyle.styled';
 
 const GoBackBtn = () => {
   const router = useRouter();
