+++ src/layouts/MainLayout.tsx
@@ -3,7 +3,7 @@ import BottomNav from '@/components/BottomNav/BottomNav';
 import styled from 'styled-components';
 import HomeHeader from '@/components/Home/HomeHeader';
 import { useSelector } from 'react-redux';
-import Header from '@/components/Header/Header';
+import Header from '@/components/Common/Header/Header';
 import { RootState } from '@/store';
 import ArchiveHeader from '@/components/Archive/ArchiveHeader';
 import SettingsHeader from '@/components/Settings/SettingsHeader';
